Fix installations with broken user face image

+++ src/installer.py
@@ -245,8 +245,9 @@ class InstallerEngine:
             os.system("rm -f /target/tmp/.passwd")
 
             # Add user's face
-            os.system("cp /tmp/live-installer-face.png /target/home/%s/.face" % setup.username)
-            self.do_run_in_chroot("chown %s:%s /home/%s/.face" % (setup.username, setup.username, setup.username))
+            if os.path.exists('/tmp/live-installer-face.png'):
+                os.system("cp /tmp/live-installer-face.png /target/home/%s/.face" % setup.username)
+                self.do_run_in_chroot("chown %s:%s /home/%s/.face" % (setup.username, setup.username, setup.username))
 
             # Make the new user the default user in KDM
             if os.path.exists('/target/etc/kde4/kdm/kdmrc'):
